Data.Thyme.Format: %s was parsed as Int, which is a Y2038 bug on 32-bit

+++ src/Data/Thyme/Format.hs
@@ -26,6 +26,7 @@ import qualified Data.Attoparsec.ByteString.Char8 as P
 import Data.Bits
 import qualified Data.ByteString.Char8 as S
 import Data.Char
+import Data.Int
 import Data.Micro
 import Data.Thyme.Calendar
 import Data.Thyme.Calendar.Internal
@@ -358,7 +359,7 @@ timeParser TimeLocale {..} = flip execStateT unixEpoch . go where
             -- UTCTime
             's' -> do
                 s <- lift (negative P.decimal)
-                _tpPOSIXTime .= fromSeconds (s :: Int)
+                _tpPOSIXTime .= fromSeconds (s :: Int64)
                 flag IsPOSIXTime .= True
                 go rspec
